+++ src/Models/PlanSubscriptionFeature.php
@@ -7,6 +7,7 @@ namespace Bpuig\Subby\Models;
 use Bpuig\Subby\Traits\HasResetDate;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
+use InvalidArgumentException;
 
 /**
  * Class PlanSubscriptionFeature
@@ -14,7 +15,6 @@ use Illuminate\Database\Eloquent\Model;
  */
 class PlanSubscriptionFeature extends Model
 {
-
     use HasResetDate;
 
     /**
@@ -22,8 +22,8 @@ class PlanSubscriptionFeature extends Model
      */
     protected $fillable = [
         'tag',
-        'subscription_id',
-        'feature_id',
+        'plan_subscription_id',
+        'plan_feature_id',
         'name',
         'description',
         'value',
@@ -37,8 +37,8 @@ class PlanSubscriptionFeature extends Model
      */
     protected $casts = [
         'tag' => 'string',
-        'subscription_id' => 'integer',
-        'feature_id' => 'integer',
+        'plan_subscription_id' => 'integer',
+        'plan_feature_id' => 'integer',
         'value' => 'string',
         'resettable_period' => 'integer',
         'resettable_interval' => 'string',
@@ -68,8 +68,8 @@ class PlanSubscriptionFeature extends Model
                 'required',
                 'max:150'
             ],
-            'subscription_id' => 'required|integer|exists:' . config('subby.tables.plan_subscriptions') . ',id',
-            'feature_id' => 'nullable|integer',
+            'plan_subscription_id' => 'required|integer|exists:' . config('subby.tables.plan_subscriptions') . ',id',
+            'plan_feature_id' => 'nullable|integer',
             'name' => 'required|string|max:150',
             'description' => 'nullable|string|max:32768',
             'value' => 'required|string',
@@ -81,23 +81,23 @@ class PlanSubscriptionFeature extends Model
 
 
     /**
-     * The feature belongs to one subscription.
+     * The subscription feature belongs to one subscription.
      *
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
     public function subscription()
     {
-        return $this->belongsTo(config('subby.models.plan_subscription'), 'subscription_id', 'id');
+        return $this->belongsTo(config('subby.models.plan_subscription'), 'plan_subscription_id', 'id');
     }
 
     /**
-     * The feature belongs to one feature.
+     * The subscription feature belongs to one plan feature.
      *
      * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
      */
     public function feature()
     {
-        return $this->belongsTo(config('subby.models.plan_feature'), 'feature_id', 'id');
+        return $this->belongsTo(config('subby.models.plan_feature'), 'plan_feature_id', 'id');
     }
 
     /**
@@ -107,9 +107,49 @@ class PlanSubscriptionFeature extends Model
      */
     public function scopeWithoutPlan(Builder $query)
     {
-        return $query->orWhereHas('feature', function (Builder $query) {
+        return $query->whereHas('feature', function (Builder $query) {
             $query->whereNull('plan_id');
         })
-            ->orWhereNull('feature_id');
+            ->orWhereNull('plan_feature_id');
+    }
+
+    /**
+     * Sync feature with subscription related plan
+     * @return $this
+     */
+    public function syncPlanSubscription()
+    {
+        $planFeature = $this->subscription->plan->getFeatureByTag($this->tag);
+        $this->syncPlanFeature($planFeature);
+
+        return $this;
+    }
+
+    /**
+     * Sync feature with related plan feature
+     * @param PlanFeature|null $planFeature
+     * @return PlanSubscriptionFeature
+     */
+    public function syncPlanFeature(PlanFeature $planFeature = null): PlanSubscriptionFeature
+    {
+        if (!$planFeature && $this->plan_feature_id) {
+            // If no Plan Feature specified, use plan in related feature (feature_id)
+            $planFeature = $this->feature;
+        } elseif (!$planFeature && !$this->plan_feature_id) {
+            // There is no way to synchronize with a plan
+            throw new InvalidArgumentException('Feature is not related to a plan.');
+        }
+
+        $this->plan_feature_id = $planFeature->id;
+        $this->name = $planFeature->name;
+        $this->description = $planFeature->description;
+        $this->value = $planFeature->value;
+        $this->resettable_period = $planFeature->resettable_period;
+        $this->resettable_interval = $planFeature->resettable_interval;
+        $this->sort_order = $planFeature->sort_order;
+
+        $this->save();
+
+        return $this;
     }
 }