+++ tests/Unit/PlanSubscriptionFeatureWithoutPlanUsageTest.php
@@ -25,7 +25,7 @@ class PlanSubscriptionFeatureWithoutPlanUsageTest extends TestCase
     public function testCannotUseFeatureExistingInCurrentRelatedPlan()
     {
         $this->expectException('Illuminate\Database\QueryException');
-        $this->expectExceptionMessage('UNIQUE constraint failed: plan_subscription_features.subscription_id, plan_subscription_features.tag');
+        $this->expectExceptionMessage('UNIQUE constraint failed: plan_subscription_features.plan_subscription_id, plan_subscription_features.tag');
         $this->testUser->subscription('main')->features()->create(['tag' => 'social_profiles', 'name' => 'Social profiles', 'value' => 10, 'sort_order' => 10]);
     }